@@ -1,40 +1,29 @@
 {
   "name": "vue-h5-template",
-  "version": "0.3.1",
-  "packageManager": "pnpm@8.1.1",
+  "type": "module",
+  "version": "0.4.0",
+  "private": true,
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview --port 4173",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "prepare": "husky install"
-  },
-  "lint-staged": {
-    "*": "eslint . --fix"
+    "preview": "vite preview",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "axios": "^1.3.4",
-    "js-cookie": "^3.0.1",
     "nprogress": "^0.2.0",
-    "pinia": "^2.0.34",
-    "vant": "^2.12.54",
-    "vue": "^2.7.14",
-    "vue-router": "^3.6.5"
+    "pinia": "^2.1.7",
+    "vant": "^4.8.5",
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
-    "@leedomjs/eslint-config": "^0.7.0",
-    "@vitejs/plugin-legacy": "^4.0.2",
-    "@vitejs/plugin-vue2": "^2.2.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.47.0",
-    "husky": "^8.0.3",
-    "lint-staged": "^14.0.1",
-    "postcss": "^8.4.21",
-    "sass": "^1.60.0",
-    "tailwindcss": "^3.3.1",
-    "terser": "^5.16.8",
-    "unplugin-vue-components": "^0.24.1",
-    "vite": "^4.2.1"
+    "@leedomjs/eslint-config": "^0.8.2",
+    "@vant/auto-import-resolver": "^1.0.2",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "eslint": "^8.57.0",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.1.6",
+    "vite-plugin-vue-devtools": "^7.0.17"
   }
 }
