@@ -1,5 +1,7 @@
 <template>
-  <div id="app">
-    <router-view />
-  </div>
+  <nav>
+    <router-link to="/">home</router-link>
+    <router-link to="/about">about</router-link>
+  </nav>
+  <RouterView />
 </template>
